Move Schedule and Event Tab to Tab Folder
