+++ app/webpacker/components/CompetitionTabs/Schedule/TimeZone.jsx
@@ -7,9 +7,9 @@ import {
   Segment,
 } from 'semantic-ui-react';
 import _ from 'lodash';
-import I18n from '../../lib/i18n';
-import { backendTimezones } from '../../lib/wca-data.js.erb';
-import { sortByOffset } from '../../lib/utils/timezone';
+import I18n from '../../../lib/i18n';
+import { backendTimezones } from '../../../lib/wca-data.js.erb';
+import { sortByOffset } from '../../../lib/utils/timezone';
 
 // Timezones that our Ruby backend knows about. They represent values that might be stored
 //   in the 'competition_venues' table.
