+++ app/webpacker/components/CompetitionTabs/Schedule/index.jsx
@@ -5,11 +5,11 @@ import TableView from './TableView';
 import TimeZoneSelector from './TimeZone';
 import VenuesAndRooms from './VenuesAndRooms';
 import ViewSelector from './ViewSelector';
-import useStoredState from '../../lib/hooks/useStoredState';
-import { earliestWithLongestTieBreaker } from '../../lib/utils/activities';
-import { getDatesBetweenInclusive } from '../../lib/utils/dates';
-import { EventSelector } from '../CompetitionsOverview/CompetitionsFilters';
-import I18n from '../../lib/i18n';
+import useStoredState from '../../../lib/hooks/useStoredState';
+import { earliestWithLongestTieBreaker } from '../../../lib/utils/activities';
+import { getDatesBetweenInclusive } from '../../../lib/utils/dates';
+import { EventSelector } from '../../CompetitionsOverview/CompetitionsFilters';
+import I18n from '../../../lib/i18n';
 
 const activeIdReducer = (state, { type, id, ids }) => {
   let newState = [...state];
